Moves filter model back to Filter container
