--- src/models/filter/constants.js
@@ -1,9 +1,0 @@
-export const SAVE_FILTER = 'models/filter/SAVE_FILTER';
-export const SAVE_FILTER_SUCCESS = 'models/filter/SAVE_FILTER_SUCCESS';
-export const SAVE_FILTER_FAILED = 'models/filter/SAVE_FILTER_FAILED';
-
-export const UPDATE_FILTER = 'models/filter/UPDATE_FILTER';
-export const UPDATE_FILTER_SUCCESS = 'models/filter/UPDATE_FILTER_SUCCESS';
-export const UPDATE_FILTER_FAILED = 'models/filter/UPDATE_FILTER_FAILED';
-
-export const CLEAR_FILTER = 'models/filter/CLEAR_FILTER';
