+++ src/signals/incident-management/containers/Filter/__tests__/actions.test.js
@@ -3,7 +3,7 @@ import { testActionCreator } from 'test/utils';
 import * as constants from '../constants';
 import * as actions from '../actions';
 
-describe('models/filter/actions', () => {
+describe('signals/incident-management/containers/Filter/actions', () => {
   it('should dispatch filterSaved action', () => {
     const foo = 'bar';
     const payload = { foo };
