+++ src/signals/incident-management/containers/Filter/actions.js
