+++ src/signals/incident-management/containers/Filter/constants.js
@@ -1,0 +1,9 @@
+export const SAVE_FILTER = 'signals/incident-management/containers/Filter/SAVE_FILTER';
+export const SAVE_FILTER_SUCCESS = 'signals/incident-management/containers/Filter/SAVE_FILTER_SUCCESS';
+export const SAVE_FILTER_FAILED = 'signals/incident-management/containers/Filter/SAVE_FILTER_FAILED';
+
+export const UPDATE_FILTER = 'signals/incident-management/containers/Filter/UPDATE_FILTER';
+export const UPDATE_FILTER_SUCCESS = 'signals/incident-management/containers/Filter/UPDATE_FILTER_SUCCESS';
+export const UPDATE_FILTER_FAILED = 'signals/incident-management/containers/Filter/UPDATE_FILTER_FAILED';
+
+export const CLEAR_FILTER = 'signals/incident-management/containers/Filter/CLEAR_FILTER';
