+++ src/signals/incident-management/containers/Filter/reducer.js
