+++ src/signals/incident-management/containers/Filter/selectors.js
@@ -1,7 +1,7 @@
 import { createSelector } from 'reselect';
 import { initialState } from './reducer';
 
-const selectFilters = (state) => state.get('filterModel') || initialState;
+const selectFilters = (state) => state.get('incidentManagementFilter') || initialState;
 
 export const makeSelectActiveFilter = createSelector(
   selectFilters,
